Remove timeout for serial port
The timeout was a dirty hack from the start.  It was long enough that
commands have finished running by the time it completes, but it wasn't
the right way to see when a command finished running.  The right way is
to look for the PDBS) prompt, which is what we do now.

This makes the Sink.set_tmpcfg method run noticeably faster.  Neat.