Added exception-raising code to fix failing tests
Added code to set_v and set_i to check if the Sink gave any error
message when running those commands, raising exceptions if it did.
This makes two of the three new tests pass.  The other one will need
support from the Sink itself, which currently doesn't care if the
voltage is over 20 V or the current is over 5 A.  That is to say that
the third test will automatically start passing if the Sink is running
firmware 1.0.1, which is yet to be released.