Fix an exception when no capabilities are read
The intended behavior for Sink.get_source_cap() when no
Source_Capabilities are read was to return an empty list.  This wasn't
done before, instead raising a ValueError when read_pdo() tried to parse
"No Source_Capabilities" as an UnknownPDO.  Now we're a bit more
cautious, and the bug is fixed.