Added calculations for PDO lists
Now there are two functions to perform calculations on PDO lists:
calculate_pdp and follows_power_rules.  calculate_pdp calculates an
estimate of the power supply's PDP, and follows_power_rules tries to
determine if the power supply follows the USB PD Power Rules.  Either
one could run into problems that cannot be detected, but for the
problems that can be detected there's a whole mound of new test code.