Added test_sink script
To replace my old Sink testing system (GNU Screen and a text file with
the commands to send to the Sink), I wrote a Python script that tests
the Sink automatically.  The only thing I have to do (assuming the test
is successful) is press Enter a few times while watching a voltmeter's
display.  A basic testing script could be useful to other people, so I'm
including it in this Python package.

I have a vague plan to integrate this with Kaylee at some point so I
don't even have to press Enter, but that will have to wait until Kaylee
has menu support.