Auto-detect device in test_sink script
If there's a single PD Buddy Sink attached, it's pretty clear which one
we want to be testing.  As such, the testing script shouldn't require
the device to be specified if it can find exactly one Sink.  This commit
does just that, and like so, I will almost never have to figure out
which device file I need when testing a Sink again.