Improve Sink testing script
This commit updates the Sink testing script to automatically handle the
connector-flipping portion of the test, which I had to perform manually
up until now.  It also shortens the procedure, and tests that the
configuration is maintained after re-connecting the device.  And now it
should be possible to make another testing procedure inheriting from
this one for the upcoming automated test jig.